--- src/redis-latency-panel/components/constants.ts
@@ -1,17 +1,0 @@
-import { FieldName } from '../types';
-
-/**
- * Table Field names
- */
-export const DisplayNameByFieldName = {
-  [FieldName.Command]: 'Command',
-  [FieldName.Calls]: 'Number of calls',
-  [FieldName.Duration]: 'Total Duration',
-  [FieldName.DurationPerCall]: 'Duration per call',
-  [FieldName.Latency]: 'Latency',
-};
-
-/**
- * Default refresh interval
- */
-export const DefaultInterval = 1000;
