+++ src/redis-latency-panel/types.ts
@@ -1,4 +1,4 @@
-import { DataQuery } from '@grafana/data';
+import { DataFrame, DataQuery, PanelProps } from '@grafana/data';
 
 /**
  * Panel Options
@@ -19,6 +19,48 @@ export interface RedisQuery extends DataQuery {
   query?: string;
 }
 
+/**
+ * Properties
+ */
+export interface Props extends PanelProps<PanelOptions> {}
+
+/**
+ * State
+ */
+export interface State {
+  /**
+   * Table Data Frame
+   *
+   * @type {DataFrame | null}
+   */
+  tableDataFrame: DataFrame | null;
+
+  /**
+   * Current Data Frame
+   *
+   * @type {DataFrame | null}
+   */
+  currentDataFrame: DataFrame | null;
+}
+
+/**
+ * Calculation
+ */
+export interface ValuesForCalculation {
+  /**
+   * Calls
+   */
+  calls: number[];
+
+  /**
+   * Duration
+   */
+  duration: number[];
+}
+
+/**
+ * Fields
+ */
 export enum FieldName {
   Command = 'Command',
   Calls = 'Calls',
@@ -26,3 +68,19 @@ export enum FieldName {
   DurationPerCall = 'Usec_per_call',
   Latency = 'Latency',
 }
+
+/**
+ * Table Field names
+ */
+export const DisplayNameByFieldName = {
+  [FieldName.Command]: 'Command',
+  [FieldName.Calls]: 'Number of calls',
+  [FieldName.Duration]: 'Total Duration',
+  [FieldName.DurationPerCall]: 'Duration per call',
+  [FieldName.Latency]: 'Latency',
+};
+
+/**
+ * Default refresh interval
+ */
+export const DefaultInterval = 1000;
